odstranenie warninigov
git-svn-id: https://fastrpc.svn.sourceforge.net/svnroot/fastrpc/fastrpc/trunk@262 b39e363e-0249-4afc-b8ac-0517a591d126

+++ src/frpchttperror.h
@@ -46,7 +46,8 @@ namespace FRPC
 class FRPC_DLLEXPORT HTTPError_t : public ProtocolError_t
 {
 public:
-    HTTPError_t(long errNum,const char *format, ...):ProtocolError_t(errNum)
+    HTTPError_t(long errNum,const char *format, ...)
+        :   ProtocolError_t(static_cast<int>(errNum))
     {
         
         // open variadic arguments
